[Refactor] 교환 요청 컨트롤러 병합
